query-reactor/pyproject.toml
[project]
name = "queryreactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.119.0",
    "langchain-openai>=1.0.0",
    "langgraph>=0.6.10",
    "openai>=2.4.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "pydantic-ai>=0.8.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "uvicorn>=0.37.0",
    "httpx>=0.27.0",
]
